tensorflow/requirements.rpi.txt
numpy
tensorflow
pillow
opencv-python
